+++ src/ext_widgets/time_clock/time_clock.c
@@ -189,10 +189,6 @@ static ret_t time_clock_destroy(widget_t* widget) {
   TKMEM_FREE(time_clock->hour_image);
   TKMEM_FREE(time_clock->minute_image);
   TKMEM_FREE(time_clock->second_image);
-  if (time_clock->timer_id != TK_INVALID_ID) {
-    timer_remove(time_clock->timer_id);
-    time_clock->timer_id = TK_INVALID_ID;
-  }
 
   return RET_OK;
 }
@@ -299,9 +295,12 @@ widget_t* time_clock_create(widget_t* parent, xy_t x, xy_t y, wh_t w, wh_t h) {
   time_clock_t* time_clock = TKMEM_ZALLOC(time_clock_t);
   widget_t* widget = WIDGET(time_clock);
   return_value_if_fail(time_clock != NULL, NULL);
-  time_clock->timer_id = timer_add(time_clock_on_timer, widget, 1000);
 
-  return widget_init(widget, parent, &s_time_clock_vtable, x, y, w, h);
+  return_value_if_fail(widget_init(widget, parent, &s_time_clock_vtable, x, y, w, h) != NULL,
+                       widget);
+  widget_add_timer(widget, time_clock_on_timer, 1000);
+
+  return widget;
 }
 
 widget_t* time_clock_cast(widget_t* widget) {