+++ tests/widget_test.cc
@@ -327,3 +327,17 @@ TEST(Widget, index_of) {
 
   widget_destroy(w);
 }
+
+static ret_t dummy_on_timer(const timer_info_t* info) {
+  return RET_REPEAT;
+}
+
+TEST(Widget, timer) {
+  widget_t* w = window_create(NULL, 0, 0, 400, 300);
+  uint32_t id = widget_add_timer(w, dummy_on_timer, 100);
+  ASSERT_EQ(timer_find(id)->id, id);
+
+  widget_destroy(w);
+
+  ASSERT_EQ(timer_find(id), (timer_info_t*)NULL);
+}